Split tests into deterministic and hypothesis-based
By doing this, we allow the deterministic tests to be easily specified by
downstream packagers, so they can avoid the problem where Hypothesis-based
tests, because they use randomization, just run too long.
